+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased][unreleased]
+
+**Added**
+
+* Themes (Thanks to [@Seth-Rothschild](https://github.com/Seth-Rothschild))
+
+**Changed**
+
+* Update to Electron v1.6.9 beta
+* Rename 'Pause' engines command to 'Suspend'
+
+**Fixed**
+
+* Fix player names not displaying in the game chooser drawer
+
 ## [Sabaki v0.30.3][v0.30.3] (2017-05-10)
 
 **Added**